Birchbit's applied patch
