+++ arc_application/tests/selenium/selenium_task_executor.py
@@ -1,0 +1,57 @@
+class SeleniumTaskExecutor:
+    """
+    Helper class for executing reusable selenium steps
+    """
+
+    __driver = None
+    __base_url = None
+
+    def __init__(self, driver, base_url):
+        """
+        Default constructor
+        :param driver: the selenium driver to be used for executing steps
+        :param base_url: the base url against which tests are to be executed
+        """
+        self.__driver = driver
+        self.__base_url = base_url
+
+    def set_driver(self, driver):
+        """
+        Method for setting the Selenium driver instance to use for testing purposes
+        :param driver: the selenium driver instance to use for testing purposes
+        """
+        self.__driver = driver
+
+    def get_driver(self):
+        """
+        Method for getting the current Selenium driver instance
+        :return: Current Selenium driver instance
+        """
+        if self.__driver is None:
+            raise Exception("A Selenium driver has not been set. Please use the set_driver method to "
+                            "configure which driver to use for tests.")
+        return self.__driver
+
+    def set_base_url(self, base_url):
+        """
+        Method for setting the base url against which tests will be executed
+        :param base_url: the base url against which tests will be executed
+        """
+        self.__base_url = base_url
+
+    def get_base_url(self):
+        """
+        Method for getting the base url against which tests will be executed
+        :return: the base url against which tests will be executed
+        """
+        if self.__base_url is None:
+            raise Exception("No base URL for tests has been set. Please use the set_base_url method"
+                            "to configure an appropriate target URL.")
+        return self.__base_url
+
+    def navigate_to_base_url(self):
+        """
+        Method for navigating to the base url of a site
+        """
+        driver = self.get_driver()
+        driver.get(self.__base_url)
