+++ arc_application/tests/test_arc_user_summary/test_helper_methods.py
@@ -1,0 +1,43 @@
+
+import unittest
+from django.test import tag
+
+
+@tag('unit')
+class ArcUserSummaryUnitTests(unittest.TestCase):
+
+    def test_count_assigned_apps(self):
+        self.skipTest('NotImplemented')
+
+    def test_get_oldest_nanny_app_id(self):
+        self.skipTest('NotImplemented')
+
+    def test_get_oldest_childminder_app_id(self):
+        self.skipTest('NotImplemented')
+
+    def test_sort_applications_by_date_submitted(self):
+        self.skipTest('NotImplemented')
+
+    def test_assign_nanny_app_to_user(self):
+        self.skipTest('NotImplemented')
+
+    def test_assign_childminder_app_to_user(self):
+        self.skipTest('NotImplemented')
+
+    def test_list_childminder_tasks_to_review(self):
+        self.skipTest('NotImplemented')
+
+    def test_release_nanny_app_to_pool(self):
+        self.skipTest('NotImplemented')
+
+    def test_release_childminder_app_to_pool(self):
+        self.skipTest('NotImplemented')
+
+    def test_get_all_table_data(self):
+        self.skipTest('NotImplemented')
+
+    def test_get_nanny_row_data(self):
+        self.skipTest('NotImplemented')
+
+    def test_get_childminder_row_data(self):
+        self.skipTest('NotImplemented')
