+++ arc_application/tests/test_migrations/0001_initial.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Generated by Django 1.11.10 on 2018-11-12 17:24
+# Generated by Django 1.11.10 on 2019-02-15 16:58
 from __future__ import unicode_literals
 
 import django.core.validators
@@ -32,19 +32,18 @@ class Migration(migrations.Migration):
                 ('dbs_certificate_number', models.CharField(blank=True, max_length=50)),
                 ('token', models.CharField(blank=True, max_length=100, null=True)),
                 ('validated', models.BooleanField(default=False)),
-                ('current_treatment', models.NullBooleanField(blank=True)),
-                ('serious_illness', models.NullBooleanField(blank=True)),
-                ('known_to_council', models.NullBooleanField(blank=True)),
-                ('children_details', models.TextField(default='', null=True)),
-                ('hospital_admission', models.NullBooleanField(blank=True)),
+                ('current_treatment', models.NullBooleanField()),
+                ('serious_illness', models.NullBooleanField()),
+                ('known_to_council', models.NullBooleanField()),
+                ('reasons_known_to_council_health_check', models.TextField(default='', null=True)),
+                ('hospital_admission', models.NullBooleanField()),
                 ('health_check_status', models.CharField(default='To do', max_length=50)),
                 ('email_resent', models.IntegerField(default=0)),
                 ('email_resent_timestamp', models.DateTimeField(blank=True, null=True)),
-                ('lived_abroad', models.NullBooleanField(blank=True)),
-                ('military_base', models.NullBooleanField(blank=True)),
-                ('capita', models.NullBooleanField(blank=True)),
-                ('on_update', models.NullBooleanField(blank=True)),
-                ('reasons_known_to_council_health_check', models.TextField(default='', null=True)),
+                ('lived_abroad', models.NullBooleanField()),
+                ('military_base', models.NullBooleanField()),
+                ('capita', models.NullBooleanField()),
+                ('on_update', models.NullBooleanField()),
             ],
             options={
                 'db_table': 'ADULT_IN_HOME',
@@ -136,9 +135,9 @@ class Migration(migrations.Migration):
                 ('publish_details', models.NullBooleanField(default=None)),
                 ('working_in_other_childminder_home', models.NullBooleanField(default=None)),
                 ('own_children', models.NullBooleanField(default=None)),
+                ('reasons_known_to_social_services', models.TextField(default='', null=True)),
                 ('known_to_social_services_pith', models.NullBooleanField(default=None)),
-                ('reasons_known_to_social_services', models.TextField(blank=True, null=True, default=None)),
-                ('reasons_known_to_social_services_pith', models.TextField(null=True, default=""))
+                ('reasons_known_to_social_services_pith', models.TextField(default='', null=True)),
             ],
             options={
                 'db_table': 'APPLICATION',
@@ -285,11 +284,11 @@ class Migration(migrations.Migration):
                 ('lived_abroad', models.NullBooleanField()),
                 ('military_base', models.NullBooleanField()),
                 ('capita', models.NullBooleanField()),
+                ('enhanced_check', models.NullBooleanField()),
                 ('on_update', models.NullBooleanField()),
-                ('application_id', models.ForeignKey(db_column='application_id', on_delete=django.db.models.deletion.CASCADE, to='arc_application.Application')),
-                ('enhanced_check', models.NullBooleanField(blank=True)),
                 ('certificate_information', models.TextField(blank=True)),
-                ('within_three_months', models.NullBooleanField(blank=True))
+                ('within_three_months', models.NullBooleanField()),
+                ('application_id', models.ForeignKey(db_column='application_id', on_delete=django.db.models.deletion.CASCADE, to='arc_application.Application')),
             ],
             options={
                 'db_table': 'CRIMINAL_RECORD_CHECK',
@@ -316,7 +315,7 @@ class Migration(migrations.Migration):
             name='HealthCheckCurrent',
             fields=[
                 ('illness_id', models.UUIDField(default=uuid.uuid4, primary_key=True, serialize=False)),
-                ('description', models.CharField(max_length=150)),
+                ('description', models.TextField()),
                 ('person_id', models.ForeignKey(blank=True, on_delete=django.db.models.deletion.CASCADE, to='arc_application.AdultInHome')),
             ],
             options={
@@ -327,7 +326,7 @@ class Migration(migrations.Migration):
             name='HealthCheckHospital',
             fields=[
                 ('illness_id', models.UUIDField(default=uuid.uuid4, primary_key=True, serialize=False)),
-                ('description', models.CharField(max_length=150)),
+                ('description', models.TextField()),
                 ('start_date', models.DateField()),
                 ('end_date', models.DateField()),
                 ('person_id', models.ForeignKey(blank=True, on_delete=django.db.models.deletion.CASCADE, to='arc_application.AdultInHome')),
@@ -342,7 +341,7 @@ class Migration(migrations.Migration):
             name='HealthCheckSerious',
             fields=[
                 ('illness_id', models.UUIDField(default=uuid.uuid4, primary_key=True, serialize=False)),
-                ('description', models.CharField(max_length=150)),
+                ('description', models.TextField()),
                 ('start_date', models.DateField()),
                 ('end_date', models.DateField()),
                 ('person_id', models.ForeignKey(blank=True, on_delete=django.db.models.deletion.CASCADE, to='arc_application.AdultInHome')),
@@ -401,7 +400,7 @@ class Migration(migrations.Migration):
                 ('town', models.CharField(blank=True, max_length=100)),
                 ('county', models.CharField(blank=True, max_length=100)),
                 ('country', models.CharField(blank=True, max_length=100)),
-                ('postcode', models.CharField(blank=True, max_length=8)),
+                ('postcode', models.CharField(blank=True, max_length=100)),
             ],
             options={
                 'db_table': 'PREVIOUS_ADDRESS',
@@ -464,6 +463,7 @@ class Migration(migrations.Migration):
             fields=[
                 ('login_id', models.UUIDField(default=uuid.uuid4, primary_key=True, serialize=False)),
                 ('email', models.CharField(blank=True, max_length=100)),
+                ('change_email', models.CharField(blank=True, max_length=100, null=True)),
                 ('mobile_number', models.CharField(blank=True, max_length=20)),
                 ('add_phone_number', models.CharField(blank=True, max_length=20)),
                 ('email_expiry_date', models.IntegerField(blank=True, null=True)),
