+++ arc_application/tests/test_migrations/0002_auto_20190211_1433.py
@@ -1,0 +1,78 @@
+# -*- coding: utf-8 -*-
+# Generated by Django 1.11.10 on 2019-02-11 14:33
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('arc_application', '0001_initial'),
+    ]
+
+    operations = [
+        migrations.RemoveField(
+            model_name='adultinhome',
+            name='children_details',
+        ),
+        migrations.RemoveField(
+            model_name='application',
+            name='own_children_not_in_home',
+        ),
+        migrations.AddField(
+            model_name='adultinhome',
+            name='reasons_known_to_council_health_check',
+            field=models.TextField(default='', null=True),
+        ),
+        migrations.AddField(
+            model_name='application',
+            name='known_to_social_services_pith',
+            field=models.NullBooleanField(default=None),
+        ),
+        migrations.AddField(
+            model_name='application',
+            name='reasons_known_to_social_services',
+            field=models.TextField(default='', null=True),
+        ),
+        migrations.AddField(
+            model_name='application',
+            name='reasons_known_to_social_services_pith',
+            field=models.TextField(default='', null=True),
+        ),
+        migrations.AddField(
+            model_name='criminalrecordcheck',
+            name='enhanced_check',
+            field=models.NullBooleanField(),
+        ),
+        migrations.AddField(
+            model_name='userdetails',
+            name='change_email',
+            field=models.CharField(blank=True, max_length=100, null=True),
+        ),
+        migrations.AlterField(
+            model_name='arccomments',
+            name='comment',
+            field=models.CharField(blank=True, max_length=500),
+        ),
+        migrations.AlterField(
+            model_name='healthcheckcurrent',
+            name='description',
+            field=models.TextField(),
+        ),
+        migrations.AlterField(
+            model_name='healthcheckhospital',
+            name='description',
+            field=models.TextField(),
+        ),
+        migrations.AlterField(
+            model_name='healthcheckserious',
+            name='description',
+            field=models.TextField(),
+        ),
+        migrations.AlterField(
+            model_name='previousaddress',
+            name='postcode',
+            field=models.CharField(blank=True, max_length=100),
+        ),
+    ]
